+++ example/lib/ui/smart_widgets/widget_one/widget_one_viewmodel.dart
@@ -1,16 +1,11 @@
 import 'package:new_architecture/app/locator.dart';
 import 'package:new_architecture/services/information_service.dart';
 import 'package:stacked/_base_viewmodels.dart';
+import 'package:stacked/_reactive_service_mixin.dart';
 
 class WidgetOneViewModel extends ReactiveViewModel {
   final InformationService _informationService = locator<InformationService>();
   int get postCount => _informationService.postCount;
-
-  WidgetOneViewModel() {
-    reactToServices([
-      _informationService,
-    ]);
-  }
 
   void updatePostCount() {
     _informationService.updatePostCount();
@@ -23,4 +18,7 @@ class WidgetOneViewModel extends ReactiveViewModel {
   Future updateStuff() {
     return Future.delayed(const Duration(seconds: 3));
   }
+
+  @override
+  List<ReactiveServiceMixin> get reactiveServices => [_informationService];
 }