+++ CHANGELOG.md
@@ -4,3 +4,9 @@ First release of the architecture containing all the initial required parts to b
 
 
 ## [1.0.1] - 2020-04-21 - Readme instruction updates
+
+## [1.0.2] - 2020-04-22 - Reactive Viewmodel updates
+
+Instead of depending on the user to call set reactive services we require an override to provide a list of services
+
+## [1.0.3] - 2020-04-22 - Adds back setBusy call to BaseViewModel